To compile in Windows VS2010
1.var assign in header
2.construct method default/delete
3.for in
4.template parameters
5.windows header
6.some warnings

+++ src/extframe.h
@@ -151,7 +151,7 @@ public:
     virtual bool process(ConnectionImpl *connection) override
     {
         // this is an exception
-        throw ProtocolException("unimplemented frame type " + std::to_string(type()));
+        throw ProtocolException("unimplemented frame type " + std::to_string((uint64_t)type()));
 
         // unreachable
         return false;
